add IHoockPosition interface for support setting priority to hooks
