+++ wicket-orientdb/src/main/java/ru/ydn/wicket/wicketorientdb/OrientDbWebApplication.java
@@ -14,6 +14,7 @@ import org.apache.wicket.markup.html.IPackageResourceGuard;
 import org.apache.wicket.markup.html.SecurePackageResourceGuard;
 import org.apache.wicket.protocol.http.WebApplication;
 
+import ru.ydn.wicket.wicketorientdb.components.IHookPosition;
 import ru.ydn.wicket.wicketorientdb.converter.ODocumentConverter;
 import ru.ydn.wicket.wicketorientdb.converter.OIdentifiableConverter;
 import ru.ydn.wicket.wicketorientdb.rest.OrientDBHttpAPIResource;
@@ -26,15 +27,12 @@ import com.orientechnologies.orient.core.db.ODatabaseInternal;
 import com.orientechnologies.orient.core.db.ODatabaseLifecycleListener;
 import com.orientechnologies.orient.core.db.document.ODatabaseDocument;
 import com.orientechnologies.orient.core.db.record.OIdentifiable;
-import com.orientechnologies.orient.core.hook.ODocumentHookAbstract;
 import com.orientechnologies.orient.core.hook.ORecordHook;
-import com.orientechnologies.orient.core.hook.ORecordHook.HOOK_POSITION;
 import com.orientechnologies.orient.core.metadata.schema.OClass;
 import com.orientechnologies.orient.core.metadata.security.ORole;
 import com.orientechnologies.orient.core.metadata.security.ORule.ResourceGeneric;
 import com.orientechnologies.orient.core.metadata.security.OSecurity;
 import com.orientechnologies.orient.core.record.impl.ODocument;
-import com.orientechnologies.orient.core.security.OSecurityManager;
 import com.orientechnologies.orient.server.OServer;
 
 /**
@@ -146,36 +144,14 @@ public abstract class OrientDbWebApplication extends AuthenticatedWebApplication
 				for (Class<? extends ORecordHook> oRecordHookClass : hooksToRegister)
 				{
 					ORecordHook hook = createHook(oRecordHookClass, iDatabase);
-					if(hook!=null) iDatabase.registerHook(hook);
+					if(hook!=null){
+						if (hook instanceof IHookPosition){
+							iDatabase.registerHook(hook,((IHookPosition) hook).getPosition());		
+						}else{
+							iDatabase.registerHook(hook);		
+						}
+					}
 				}
-				//strange workaround to support changing system users passwords in web interface
-				iDatabase.registerHook(new ODocumentHookAbstract(){
-					@Override
-					public RESULT onRecordBeforeUpdate(final ODocument iDocument) {
-						String name = iDocument.field("name");
-						String password = iDocument.field("password");
-						if (password.startsWith(OSecurityManager.ALGORITHM_PREFIX)){
-							return RESULT.RECORD_NOT_CHANGED;
-						}
-						if (	orientDbSettings.getDBInstallatorUserName()!=null && 
-								orientDbSettings.getDBInstallatorUserName().equals(name)){
-							orientDbSettings.setDBInstallatorUserPassword(password);
-						}
-						if (orientDbSettings.getDBUserName()!=null && 
-								orientDbSettings.getDBUserName().equals(name)){
-							orientDbSettings.setDBUserPassword(password);
-						}
-						OrientDbWebSession session = OrientDbWebSession.get();
-						if (session.getUsername()!=null && session.getUsername().equals(name)){
-							session.setUser(name, password);
-						}
-						return RESULT.RECORD_NOT_CHANGED;
-					}
-					@Override
-					public DISTRIBUTED_EXECUTION_MODE getDistributedExecutionMode() {
-						return DISTRIBUTED_EXECUTION_MODE.TARGET_NODE;
-					}
-				}.setIncludeClasses("OUser"),HOOK_POSITION.FIRST);
 			}
 			
 			@Override
@@ -211,6 +187,9 @@ public abstract class OrientDbWebApplication extends AuthenticatedWebApplication
 				Orient.instance().shutdown();
 			}
 		});
+		
+		//strange workaround to support changing system users passwords in web interface
+		getOrientDbSettings().getORecordHooks().add(OUserCatchPasswordHook.class);
 	}
 	
 	protected TransactionRequestCycleListener newTransactionRequestCycleListener()
